add rabbitmq consumer and posgres connection fixed
